requirements.txt
requests==2.31.0
openai>=1.100.0
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.119.0
python-dotenv==1.0.1
schedule==1.2.0
pytz==2024.1
tenacity==8.2.3
psutil==5.9.8
beautifulsoup4==4.12.2
